+++ src/elements/map-view/config.js
@@ -1,3 +1,4 @@
 export const elementName = "map-view";
 export const defaultMapType = "";
 export const defaultViewProjection = 'EPSG:3857';
+export const defaultCenter = [0, 0];